+++ server/src/main/java/com/linbit/linstor/LinStorModule.java
@@ -1,5 +1,8 @@
 package com.linbit.linstor;
 
+import com.linbit.linstor.core.LinStor;
+import com.linbit.utils.MathUtils;
+
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -8,6 +11,12 @@ import reactor.core.scheduler.Schedulers;
 
 public class LinStorModule extends AbstractModule
 {
+    // Minimum number of worker threads
+    private static final int MIN_THREAD_SIZE = 4;
+
+    // Maximum number of worker threads
+    private static final int MAX_THREAD_SIZE = 16;
+
     @Override
     protected void configure()
     {
@@ -17,6 +26,7 @@ public class LinStorModule extends AbstractModule
     @Singleton
     public Scheduler mainWorkerPoolScheduler()
     {
-        return Schedulers.newParallel("MainWorkerPool");
+        int thrCount = MathUtils.bounds(MIN_THREAD_SIZE, LinStor.CPU_COUNT, MAX_THREAD_SIZE);
+        return Schedulers.newParallel("MainWorkerPool", thrCount);
     }
 }