+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clone: fail clone if the source resource isn't UpToDate or has skip-disk
 - Memory: Limit satellite to 2G max memory (-Xmx2G) and controller to 8G
 - JVM: will now crash on OutOfMemoryError
+- Limit reactor worker threads to 16
 
 ### Fixed